refactor Toolbar from Table
