+++ src/Components/MUARolesTable.js
@@ -1,21 +1,8 @@
 import React from 'react';
 import {
   Title,
-  Toolbar,
-  ToolbarContent,
-  ToolbarItem,
-  Select,
-  Dropdown,
-  SelectOption,
-  DropdownItem,
-  DropdownToggle,
-  InputGroup,
-  TextInput,
   Button,
   Pagination,
-  ChipGroup,
-  Chip,
-  DropdownToggleCheckbox,
   Tooltip,
   EmptyState,
   EmptyStateIcon,
@@ -29,14 +16,13 @@ import {
   Th,
   Td,
 } from '@patternfly/react-table';
-import FilterIcon from '@patternfly/react-icons/dist/js/icons/filter-icon';
 import SearchIcon from '@patternfly/react-icons/dist/js/icons/search-icon';
-import { capitalize } from '@patternfly/react-core/dist/esm/helpers/util';
 import { css } from '@patternfly/react-styles';
 import { useDispatch } from 'react-redux';
 import { addNotification } from '@redhat-cloud-services/frontend-components-notifications/redux';
 import PropTypes from 'prop-types';
 import apiInstance from '../Helpers/apiInstance';
+import RoleToolbar from './RoleToolbar';
 
 let rolesCache = [];
 let applicationsCache = [];
@@ -49,43 +35,43 @@ const MUARolesTable = ({
   const columns = ['Role name', 'Role description', 'Permissions'];
   const [rows, setRows] = React.useState(Array.from(rolesCache));
   const [applications, setApplications] = React.useState(applicationsCache);
+  const [appSelections, setAppSelections] = React.useState([]);
   React.useEffect(() => {
     if (rolesCache.length === 0 || applicationsCache.length === 0) {
-      apiInstance
-        .get(
-          `${API_BASE}/roles/?limit=9999&order_by=display_name&add_fields=groups_in_count`,
-          { headers: { Accept: 'application/json' } }
-        )
-        .then(({ data }) => {
-          data.forEach((role) => {
-            role.isExpanded = false;
-            role.permissions = role.accessCount;
-          });
-          rolesCache = data.map((role) => Object.assign({}, role));
-          setRows(data);
+      apiInstance.get(
+        `${API_BASE}/roles/?limit=9999&order_by=display_name&add_fields=groups_in_count`,
+        { headers: { Accept: 'application/json' } }
+      )
+      .then(({ data }) => {
+        data.forEach((role) => {
+          role.isExpanded = false;
+          role.permissions = role.accessCount;
+        });
+        rolesCache = data.map((role) => Object.assign({}, role));
+        setRows(data);
 
-          // Build application filter from data
-          const apps = Array.from(
-            data
-              .map((role) => role.applications)
-              .flat()
-              .reduce((acc, cur) => {
-                acc.add(cur);
-                return acc;
-              }, new Set())
-          ).sort();
-          applicationsCache = apps;
-          setApplications(apps);
-        })
-        .catch((err) =>
-          dispatch(
-            addNotification({
-              variant: 'danger',
-              title: 'Could not fetch roles list',
-              description: err.message,
-            })
-          )
-        );
+        // Build application filter from data
+        const apps = Array.from(
+          data
+            .map((role) => role.applications)
+            .flat()
+            .reduce((acc, cur) => {
+              acc.add(cur);
+              return acc;
+            }, new Set())
+        ).sort();
+        applicationsCache = apps;
+        setApplications(apps);
+      })
+      .catch((err) =>
+        dispatch(
+          addNotification({
+            variant: 'danger',
+            title: 'Could not fetch roles list',
+            description: err.message,
+          })
+        )
+      );
     }
   }, []);
 
@@ -97,15 +83,8 @@ const MUARolesTable = ({
     setActiveSortDirection(direction);
   };
 
-  // Filtering
-  const [isDropdownOpen, setIsDropdownOpen] = React.useState(false);
-  const [filterColumn, setFilterColumn] = React.useState(columns[0]);
-  const [isSelectOpen, setIsSelectOpen] = React.useState(false);
-  const [appSelections, setAppSelections] = React.useState([]);
   const [nameFilter, setNameFilter] = React.useState('');
   const hasFilters = appSelections.length > 0 || nameFilter;
-  const selectLabelId = 'filter-application';
-  const selectPlaceholder = 'Filter by application';
 
   const selectedNames = selectedRoles.map((role) => role.display_name);
   const filteredRows = rows
@@ -161,7 +140,6 @@ const MUARolesTable = ({
     .slice((page - 1) * perPage, page * perPage);
 
   // Selecting
-  const [isBulkSelectOpen, setIsBulkSelectOpen] = React.useState(false);
   const anySelected = selectedRoles.length > 0;
   const someChecked = anySelected ? null : false;
   const isChecked =
@@ -177,14 +155,6 @@ const MUARolesTable = ({
     }
   };
 
-  const onSelectAll = (_ev, isSelected) => {
-    if (isSelected) {
-      setSelectedRoles(filteredRows.map((row) => row.display_name));
-    } else {
-      setSelectedRoles([]);
-    }
-  };
-
   const clearFiltersButton = (
     <Button
       variant="link"
@@ -196,154 +166,26 @@ const MUARolesTable = ({
       Clear filters
     </Button>
   );
-  const roleToolbar = isReadOnly ? null : (
-    <Toolbar id="access-requests-roles-table-toolbar">
-      <ToolbarContent>
-        <ToolbarItem>
-          <Dropdown
-            onSelect={() => setIsBulkSelectOpen(!isBulkSelectOpen)}
-            position="left"
-            toggle={
-              <DropdownToggle
-                splitButtonItems={[
-                  <DropdownToggleCheckbox
-                    key="a"
-                    id="example-checkbox-2"
-                    aria-label={anySelected ? 'Deselect all' : 'Select all'}
-                    isChecked={isChecked}
-                    onClick={() => onSelectAll(null, !anySelected)}
-                  />,
-                ]}
-                onToggle={(isOpen) => setIsBulkSelectOpen(isOpen)}
-                isDisabled={rows.length === 0}
-              >
-                {selectedRoles.length !== 0 && (
-                  <React.Fragment>
-                    {selectedRoles.length} selected
-                  </React.Fragment>
-                )}
-              </DropdownToggle>
-            }
-            isOpen={isBulkSelectOpen}
-            dropdownItems={[
-              <DropdownItem key="0" onClick={() => onSelectAll(null, false)}>
-                Select none (0 items)
-              </DropdownItem>,
-              <DropdownItem
-                key="1"
-                onClick={() =>
-                  setSelectedRoles(
-                    selectedRoles.concat(pagedRows.map((r) => r.display_name))
-                  )
-                }
-              >
-                Select page ({Math.min(pagedRows.length, perPage)} items)
-              </DropdownItem>,
-              <DropdownItem key="2" onClick={() => onSelectAll(null, true)}>
-                Select all ({filteredRows.length} items)
-              </DropdownItem>,
-            ]}
-          />
-        </ToolbarItem>
-        <ToolbarItem>
-          <InputGroup>
-            <Dropdown
-              isOpen={isDropdownOpen}
-              onSelect={(ev) => {
-                setIsDropdownOpen(false);
-                setFilterColumn(ev.target.value);
-                setIsSelectOpen(false);
-              }}
-              toggle={
-                <DropdownToggle
-                  onToggle={(isOpen) => setIsDropdownOpen(isOpen)}
-                >
-                  <FilterIcon /> {filterColumn}
-                </DropdownToggle>
-              }
-              dropdownItems={['Role name', 'Application'].map((colName) => (
-                // Filterable columns are RequestID, AccountID, and Status
-                <DropdownItem key={colName} value={colName} component="button">
-                  {capitalize(colName)}
-                </DropdownItem>
-              ))}
-            />
-            {filterColumn === 'Application' ? (
-              <React.Fragment>
-                <span id={selectLabelId} hidden>
-                  {selectPlaceholder}
-                </span>
-                <Select
-                  aria-labelledby={selectLabelId}
-                  variant="checkbox"
-                  aria-label="Select applications"
-                  onToggle={(isOpen) => setIsSelectOpen(isOpen)}
-                  onSelect={(_ev, selection) => {
-                    if (appSelections.includes(selection)) {
-                      setAppSelections(
-                        appSelections.filter((s) => s !== selection)
-                      );
-                    } else {
-                      setAppSelections([...appSelections, selection]);
-                    }
-                  }}
-                  isOpen={isSelectOpen}
-                  selections={appSelections}
-                  isCheckboxSelectionBadgeHidden
-                  placeholderText={selectPlaceholder}
-                  style={{ maxHeight: '400px', overflowY: 'auto' }}
-                >
-                  {applications.map((app) => (
-                    <SelectOption key={app} value={app}>
-                      {capitalize(app.replace(/-/g, ' '))}
-                    </SelectOption>
-                  ))}
-                </Select>
-              </React.Fragment>
-            ) : (
-              <TextInput
-                name="rolesSearch"
-                id="rolesSearch"
-                type="search"
-                iconVariant="search"
-                aria-label="Search input"
-                placeholder="Filter by role name"
-                value={nameFilter}
-                onChange={(val) => setNameFilter(val)}
-              />
-            )}
-          </InputGroup>
-        </ToolbarItem>
-        <ToolbarItem variant="pagination" align={{ default: 'alignRight' }}>
-          <AccessRequestsPagination id="top" />
-        </ToolbarItem>
-      </ToolbarContent>
-      {hasFilters && (
-        <ToolbarContent>
-          {nameFilter && (
-            <ChipGroup categoryName="Role name">
-              <Chip onClick={() => setNameFilter('')}>{nameFilter}</Chip>
-            </ChipGroup>
-          )}
-          {appSelections.length > 0 && (
-            <ChipGroup categoryName="Status">
-              {appSelections.map((status) => (
-                <Chip
-                  key={status}
-                  onClick={() =>
-                    setAppSelections(appSelections.filter((s) => s !== status))
-                  }
-                >
-                  {status}
-                </Chip>
-              ))}
-            </ChipGroup>
-          )}
-          {clearFiltersButton}
-        </ToolbarContent>
-      )}
-    </Toolbar>
-  );
+  
+  const roleToolbar = isReadOnly
+                        ? null
+                        : <RoleToolbar 
+                            selectedRoles={selectedRoles}
+                            setSelectedRoles={setSelectedRoles}
+                            isChecked={isChecked}
+                            appSelections={appSelections}
+                            setAppSelections={setAppSelections}
+                            columns={columns}
+                            rows={rows}
+                            filteredRows={filteredRows}
+                            pagedRows={pagedRows}
+                            clearFiltersButton={clearFiltersButton}
+                            perPage={perPage}
+                            nameFilter={nameFilter}
+                            setNameFilter={setNameFilter}
+                            AccessRequestsPagination={AccessRequestsPagination}
+                            applications={applications}
+                          />;
 
   const expandedColumns = ['Application', 'Resource type', 'Operation'];
   const dispatch = useDispatch();
@@ -386,7 +228,7 @@ const MUARolesTable = ({
               columnIndex: 'name',
             }}
           >
-            {columns[0]}
+          {columns[0]}
           </Th>
           <Th
             width={50}
@@ -399,7 +241,7 @@ const MUARolesTable = ({
               columnIndex: 'description',
             }}
           >
-            {columns[1]}
+          {columns[1]}
           </Th>
           <Th
             width={10}
@@ -413,7 +255,7 @@ const MUARolesTable = ({
             }}
             modifier="nowrap"
           >
-            {columns[2]}
+          {columns[2]}
           </Th>
         </Tr>
       </Thead>
@@ -428,7 +270,7 @@ const MUARolesTable = ({
                     style={{ height: '22px' }}
                     className="ins-c-skeleton ins-c-skeleton__md"
                   >
-                    {' '}
+                  {' '}
                   </div>
                 </Td>
               ))}
@@ -518,21 +360,23 @@ const MUARolesTable = ({
         </Tbody>
       ))}
       {pagedRows.length === 0 && hasFilters && (
-        <Tr>
-          <Td colSpan={columns.length}>
-            <EmptyState variant="small">
-              <EmptyStateIcon icon={SearchIcon} />
-              <Title headingLevel="h2" size="lg">
-                No matching requests found
-              </Title>
-              <EmptyStateBody>
-                No results match the filter criteria. Remove all filters or
-                clear all filters to show results.
-              </EmptyStateBody>
-              {clearFiltersButton}
-            </EmptyState>
-          </Td>
-        </Tr>
+        <Tbody>
+          <Tr>
+            <Td colSpan={columns.length}>
+              <EmptyState variant="small">
+                <EmptyStateIcon icon={SearchIcon} />
+                <Title headingLevel="h2" size="lg">
+                  No matching requests found
+                </Title>
+                <EmptyStateBody>
+                  No results match the filter criteria. Remove all filters or
+                  clear all filters to show results.
+                </EmptyStateBody>
+                {clearFiltersButton}
+              </EmptyState>
+            </Td>
+          </Tr>
+        </Tbody>
       )}
     </TableComposable>
   );
